Cloud Functions/function-2-readme/requirements.txt
functions-framework==3.5.0
google-cloud-aiplatform >= 1.31.0
google-cloud-logging